Replace Router#observeHistory() (#37)
The history manager can now be provided when creating the router. It is
still optional, to allow for routers to be created that handle
server-side requests in Node.js.

Similarly a context object can be provided, or a factory function which
returns a context object for each request.

The new Router#start() causes the router to observe the history manager
and dispatch requests in response. By default it performs an immediate
dispatch based on the current history value. Calling start() on a router
that does not have a history manager is a noop. start() can only be
called once.