Enable strict null checks (#54)
* Enable strictNullChecks

* Work around poor Immutable typings

* Set default projector tagName upon creation

Move the default from attachment to creation. This helps with strict
null checks since the tagName is now always set to a string.

* Initial projector state properties shouldn't be optional

These properties are always set, so mark them as required to avoid
strict null check violations.

* Initialize noop projector data properties

These properties are set to noop values after detaching, so set them
when the data is created to avoid strict null check violations.

* Ensure Projector#attach() returns a promise for a handle

attachPromise is optional, so explicitly fall back to a promise for the noop handle.

* Explicitly mark Dijit properties as potentially being undefined

The tests already deal with this.

* Properly handle resolving an undefined Dijit constructor

The tests already account for the constructor being undefined.

* More reliable default params for Dijit constructor calls

Default to an empty object when calling the constructor. This works
better with strict null checks and avoids needing to set defaults in
the two places where params can be assigned to.

Add a test to ensure params defaults to an empty object.

* Ensure resize listeners return booleans

* Export a stubbed projector in non-browser envs

Projectors cannot be created outside of browser environments. Ensure
that a default projector can always be exported, even if it can't do
anything.

Exporting a null value means all code that binds against the default
projector needs to cope with there possibly not being one. Rather than
having design time violations of the strict null check kind, we'll now
have runtime errors when the stubbed projector is used.

* Return default resize panel width in getter

Return the default width if it hasn't been set, or is the empty string.
This is preferable over falling back to the default width when building
the node attributes, and helps avoid strict null check violations.

* Fix compile errors in examples

* Remove unnecessary union for ParentListMixin this type

* Return empty array when there are no children nodes

* Type RenderMixin initialize arguments

* Initialize StatefulChildrenMixin cache

Avoid strict null check violations by initializing as part of the
destructure, then assign if the internal state didn't yet have a cache.

* Don't create List with undefined argument

* Initialize StatefulListenersMixin cache

Avoid strict null check violations by initializing as part of the
destructure, then assign if the internal state didn't yet have a cache.

* Explicit undefined when reducing in StatefulListenersMixin

The reducer returns undefined when the listeners were resolved
synchronously. Mark it as undefined to avoid strict null check
violations.

* Refactor listener resolution

Now without needing type casts, nor strict null check violations.

* Compare optional boolean to true in TabbedMixin

* Guard against optional label state being undefined

* Remove unnecessary type union

* Guard against classes property being undefined

(Not that I think it ever is…)

* VNodeListeners index type must allow undefined values

Otherwise the optional listeners can't be assigned, or something.

* Stop casting null to a VNodeListeners

Casting null is not allowed when strict null checks is enabled. Instead
provide a empty, frozen object.

* Avoid calling indexOf with an undefined value

* Correctly reset parent references to null

* Fix initialBaseUrl typing

* Fix type warnings in Dijit support class

* Ignore undefined type errors in createDijit test

* Ignore undefined type errors in createList test

* Fix on('valuechange') typing in FormFieldMixin

* Fix strict null check violations in StatefulChildrenMixin test

* Fix strict null check violations in StatefulListenersMixin test

* Fix strict null check violations in TabbedMixin test

* Bump minimal versions of dojo-core and dojo-shim peer dependencies

* Require particular maquette versions

Typings have changed in 2.3.7. Widgets now depends on at least that
version. Cap at 2.4.1 to avoid accidentally pulling in newer versions
with similar problems.

* Silence null checks in tests

Silence violations that came up after upgrading Maquette.