Implement link generation (#53)
* Create explicit route Handler type

* Create explicit type for known search params

* Include path and raw values in Route's Selection

We need to track these values so the Router can generate links for
currently selected routes.

Refactor Route#select() to remove deeply nested return statements and
avoid repetition in constructing the Selection object.

* Ensure routes are only appended once

Generating links requires the router to walk the route hierarchy back to
the router itself. This means routes should only be appended to other
routes or the router once.

Add a parent accessor to routes. This can be used both to see whether
the route has previously been appended to another route, and to
walk the hierarchy.

* Add prefix() method to history managers

This allows the router to generate links that work in the browser. For
hash-based history managers this prefixes with #, for state-based
history managers the prefix includes any base path.

* Track selected routes if dispatched from start()

* Expose deconstructed path on Route

The router needs access to the path in order to generate links. Ensure
the path is now deeply frozen since it must not be modified after routes
are created.

* Track whether deconstructed path started with a slash

The slash needs to be included in generated links.

* Implement Router#link()

Fixes #46.