I18nMixin: return blank locale bundle by default (#876)
Update `I18nMixin#localizeBundle` to return an object with `format`,
`isPlaceholder`, and `messages` properties, instead of a single
object containing both the messages and the `format` method. The
new `isPlaceholder` property indicates whether the returned messages
are just the (blank or default) placeholders or the actual
locale-specific messages.

By default, return `I18nMixin#localizeBundle` a blank locale bundle
while waiting for locale-specific messages to load. If `true` is passed as
the second argument, return the default messages instead of a blank
bundle.